creating item-overview component- show the title, id, lables, rating, price, etc... and other informations based on mockup- creating description, detail, specification tabs
