+++ src/app/app-routing.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { ContentComponent } from './content/content.component';
-import { ItemComponent } from './content/item/item.component';
+import { ItemOverviewComponent } from './content/item-overview/item-overview.component';
 import { UserProfileComponent } from './content/user-profile/user-profile.component';
 
 const routes: Routes = [
   { path: '', component: ContentComponent },
   { path: 'user-profile', component: UserProfileComponent },
-  { path: 'item/:id', component: ItemComponent }
+  { path: 'item/:id', component: ItemOverviewComponent }
 ];
 
 @NgModule({
