+++ src/app/app.component.html
@@ -1,3 +1,4 @@
 <app-header></app-header>
+<!--<app-content></app-content>-->
 <router-outlet></router-outlet>
 <app-footer></app-footer>
