+++ src/app/content/content.component.html
@@ -1,24 +1,9 @@
-<div class="content">
-    <mat-tab-group>
-        <mat-tab label="Today's Sales">
-            <mat-grid-list cols="5" rowHeight="3:2">
-                <mat-grid-tile *ngFor="let item of items">
-                    <mat-list class="item-list" routerLink="/item/{{ item.id }}" class="item-widget">
-                        <mat-list-item class="item-info"><b>{{ item.name }}</b></mat-list-item>
-                        <mat-list-item><img class="item-photo" src='{{ item.photos[0] }}' /></mat-list-item>
-                        <mat-list-item class="item-info">Rating: {{ item.avgRating }}/5 ({{ item.nrOfRatings }})</mat-list-item>
-                        <mat-list-item class="item-info">
-                            <mat-chip-list>
-                                <mat-chip disabled *ngFor="let label of item.labels">
-                                    {{ label.key ? label.key + ' : ' : '' }}{{ label.value }}
-                                </mat-chip>
-                            </mat-chip-list>
-                        </mat-list-item>
-                    </mat-list>
-                </mat-grid-tile>
-            </mat-grid-list>
-        </mat-tab>
-        <mat-tab label="Vouchers">TBD</mat-tab>
-        <mat-tab label="Brewing Guide">TBD</mat-tab>
-    </mat-tab-group>
-</div>
+<app-item></app-item>
+<!--<div class="content">-->
+<!--    <mat-tab-group>-->
+<!--        <mat-tab label="Today's Sales"></mat-tab>-->
+<!--        <mat-tab label="Vouchers">TBD</mat-tab>-->
+<!--        <mat-tab label="Brewing Guide">TBD</mat-tab>-->
+<!--    </mat-tab-group>-->
+<!--</div>-->
+
