+++ src/app/content/item/item.component.ts
@@ -1,25 +1,43 @@
-import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-import { ApiService } from 'src/api/api.service';
+import {Component, OnInit, ViewEncapsulation} from '@angular/core';
+import {MatDialog, MatDialogRef} from '@angular/material/dialog';
+import {ApiService} from '../../../api/api.service';
+import {ProgressSpinnerDialogComponent} from '../../widgets/progress-spinner-dialog/progress-spinner-dialog.component';
 
 @Component({
   selector: 'app-item',
   templateUrl: './item.component.html',
-  styleUrls: ['./item.component.scss']
+  styleUrls: ['./item.component.scss'],
+  encapsulation: ViewEncapsulation.None
 })
 export class ItemComponent implements OnInit {
+  dialogRef!: MatDialogRef<ProgressSpinnerDialogComponent>;
+  items: any;
 
-  itemDetails: any;
-
-  constructor(private service: ApiService, private route: ActivatedRoute) { }
-
-  ngOnInit(): void {
-    let itemId = this.route.snapshot.paramMap.get('id');
-    if (itemId && itemId != '') {
-      this.service.getItemDetails(itemId).subscribe((data: any) => {
-        this.itemDetails = data;
-      });
-    }
+  constructor(private dialog: MatDialog, private apiService: ApiService) {
+    this.showSpinner();
   }
 
+  ngOnInit(): void {
+    this.getItems(0, 10);
+  }
+
+  getItems(page: number, size: number) {
+    this.apiService.getItems(page, size)
+      .subscribe((data: any) => {
+        this.items = data.content;
+
+        this.hideSpinner();
+      });
+  }
+
+  showSpinner() {
+    this.dialogRef = this.dialog.open(ProgressSpinnerDialogComponent, {
+      panelClass: 'transparent',
+      disableClose: true,
+    });
+  }
+
+  hideSpinner() {
+    this.dialogRef.close();
+  }
 }
